+++ pytest-selenium/pages/content.py
@@ -9,7 +9,6 @@ from regions.base import Region
 
 class Content(Page):
     URL_TEMPLATE = "/books/{book_slug}/pages/{page_slug}"
-    _title_locator = (By.TAG_NAME, "title")
     _body_locator = (By.TAG_NAME, "body")
     _main_content_locator = (By.CSS_SELECTOR, "h1")
     _next_locator = (By.CSS_SELECTOR, "[aria-label='Next Page']")
@@ -18,14 +17,6 @@ class Content(Page):
     @property
     def loaded(self):
         return self.find_element(*self._body_locator).get_attribute("data-rex-loaded")
-
-    @property
-    def title(self):
-        return self.find_element(*self._title_locator)
-
-    @property
-    def title_before_click(self):
-        return self.title.get_attribute("innerHTML")
 
     @property
     def next_link(self):
